+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Unreleased:
+v0.8.1 - 2022-11-30:
   * Add `Buffer::new_preallocated_owned(...)` to transfer ownership of a pre-allocated buffer to Aravis.
   * Deprecate `Buffer::new_preallocated(...)` in favor of `Buffer::new_preallocated_borrowed`.